+++ src/shared/components/ui/progress.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { Progress as ProgressPrimitive } from "radix-ui";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 function Progress({
 	className,
