+++ src/shared/components/ui/scroll-area.tsx
@@ -1,7 +1,7 @@
 import * as React from "react";
 import { ScrollArea as ScrollAreaPrimitive } from "radix-ui";
 
-import { cn } from "@/lib/utils";
+import { cn } from "@/shared/lib/utils";
 
 function ScrollArea({
 	className,
